[ACS-5051] Playwright files extracted
