+++ e2e-playwright/.eslintrc.json
@@ -1,0 +1,26 @@
+{
+  "extends": "../.eslintrc.json",
+  "ignorePatterns": [
+    "!**/*"
+  ],
+  "overrides": [
+    {
+      "files": [
+        "*.ts"
+      ],
+      "parserOptions": {
+        "project": [
+          "e2e-playwright/tsconfig.e2e.json"
+        ],
+        "createDefaultProgram": true
+      },
+      "plugins": [
+        "rxjs",
+        "unicorn"
+      ],
+      "rules": {
+        "@typescript-eslint/no-floating-promises": "off"
+      }
+    }
+  ]
+}
