+++ e2e-playwright/shared/base-config/index.ts
