+++ e2e-playwright/shared/page-objects/components/spinner.component.ts
