+++ e2e-playwright/shared/page-objects/global-variables.ts
