+++ e2e-playwright/test/page-objects/components/conditions.component.ts
