+++ e2e-playwright/test/page-objects/components/manage-rules-dialog.component.ts
