+++ e2e-playwright/test/tests/folder-rules/conditions.spec.ts
@@ -22,13 +22,15 @@
  * from Hyland Software. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { test } from '../fixtures/page-initialization';
-import { ActionType } from '../page-objects/components/actions-dropdown.component';
+import { test } from '../../fixtures/page-initialization';
+import { ActionType } from '../../page-objects/components/actions-dropdown.component';
+import { Comparator, Field } from '../../page-objects/components/conditions.component';
 import { expect } from '@playwright/test';
 
-test.describe('Folder Rules Actions', () => {
+test.describe('Folder Rules Conditions', () => {
   const randomFolderName = `playwright-folder-${(Math.random() + 1).toString(36).substring(6)}`;
   const randomRuleName = `playwright-rule-${(Math.random() + 1).toString(36).substring(6)}`;
+  const specialChars = '!@£$%^&*()~#/';
 
   let folderId: string;
 
@@ -38,18 +40,19 @@ test.describe('Folder Rules Actions', () => {
   });
 
   test.beforeEach(async ({ personalFiles }) => {
-    await personalFiles.navigate({ waitUntil: 'domcontentloaded' });
+    await personalFiles.navigate();
   });
 
   test.afterAll(async ({ apiClient }) => {
     await apiClient.nodes.deleteNode(folderId);
   });
 
-  test('Create a rule with actions', async ({ personalFiles, nodesPage }) => {
+  test('Create a rule with condition', async ({ personalFiles, nodesPage }) => {
     await personalFiles.dataTable.performActionFromExpandableMenu(randomFolderName, 'Manage rules');
 
     await nodesPage.toolbar.clickCreateRuleButton();
     await nodesPage.manageRulesDialog.typeRuleName(randomRuleName);
+    await nodesPage.conditionsDropdown.addCondition(Field.Size, Comparator.Equals, specialChars);
     await nodesPage.actionsDropdown.selectAction(ActionType.IncrementCounter);
     await nodesPage.manageRulesDialog.createRuleButton.click();
