+++ tsconfig.json
@@ -31,7 +31,7 @@
       "@alfresco/aca-content/about": ["projects/aca-content/about/src/public-api.ts"],
       "@alfresco/aca-viewer": ["projects/aca-viewer/src/public-api.ts"],
       "@alfresco/aca-preview": ["projects/aca-preview/src/public-api.ts"],
-      "@alfresco/playwright-shared": ["e2e/playwright/shared/index.ts"],
+      "@alfresco/playwright-shared": ["e2e-playwright/shared/index.ts"],
       "package.json": ["package.json"]
     }
   },
